Inimigos e pulo
Implementei o pulo através da equação de torricelli

Implementei as funcionalidades dos inimigos

+++ game/Dragao.h
@@ -11,22 +11,16 @@ private:
 	//Quando o Jogador encosta nele ele é empurrado para o lado e para cima
 	//NAO SEGUE O JOGADOR
 
-	int tempo_de_ataque;
+	float tempo_de_ataque;
 
 public:
 	Dragao();
 	~Dragao();
 
-	const int getTempoAtaque() const;
+	const float getTempoAtaque() const;
 
 	virtual void Mover();
 	virtual void Executar();
-
-
-
-
-
-
 
 };
