+++ game/Obstaculo.cpp
@@ -12,7 +12,7 @@ void Obstaculo::anulaGravidade()
 {
 	if (gravidade) {
 
-		double vetor_contrario = speed.y * -1.0;
+		float vetor_contrario = speed.y * -1.0;
 		corpo.move(Vector2f(0.0f, vetor_contrario));
 	}
